moved file tree stuff to web interface
